requirements.txt
streamlit-carousel==1.0.0
streamlit-option-menu==0.3.13
matplotlib==3.9.0
wordcloud==1.9.3

